rep bins as uint64

+++ src/bow_computeImageRep.m
@@ -12,4 +12,4 @@ catch
 end
 I = single(I);
 [f, d] = vl_sift(I);
-bins = double(vl_kdtreequery(model.kdtree, model.vocab, double(d)));
+bins = uint64(vl_kdtreequery(model.kdtree, model.vocab, double(d)));
